// src/entities/Profile/model/types/profile.ts
import {Currency} from "entities/Currency";
import {Country} from "entities/Country";

export interface Profile {
    firstName?: string;
    lastName?: string;
    age?: number;
    currency?: Currency;
    country?: Country;
    city?: string;
    userName?: string;
    avatar?: string;
}

export type ProfileKey = keyof Profile;

export interface ProfileSchema {
    data?: Profile;
    form?: Profile;
    isLoading: boolean;
    error?: string;
    readonly: boolean
}